feat(linter): migrate eslint executor to devkit
