+++ packages/linter/builders.json
@@ -1,14 +1,25 @@
 {
-  "$schema": "@angular-devkit/architect/src/builders-schema.json",
   "builders": {
     "lint": {
-      "implementation": "./src/builders/lint/lint.impl",
-      "schema": "./src/builders/lint/schema.json",
+      "implementation": "./src/executors/lint/compat",
+      "schema": "./src/executors/lint/schema.json",
       "description": "**[DEPRECATED]**: Please use the eslint builder instead, an automated migration was provided in v10.3.0"
     },
     "eslint": {
-      "implementation": "./src/builders/eslint/lint.impl",
-      "schema": "./src/builders/eslint/schema.json",
+      "implementation": "./src/executors/eslint/compat",
+      "schema": "./src/executors/eslint/schema.json",
+      "description": "Run ESLint on a project"
+    }
+  },
+  "executors": {
+    "lint": {
+      "implementation": "./src/executors/lint/lint.impl",
+      "schema": "./src/executors/lint/schema.json",
+      "description": "**[DEPRECATED]**: Please use the eslint builder instead, an automated migration was provided in v10.3.0"
+    },
+    "eslint": {
+      "implementation": "./src/executors/eslint/lint.impl",
+      "schema": "./src/executors/eslint/schema.json",
       "description": "Run ESLint on a project"
     }
   }
