+++ packages/linter/src/executors/eslint/compat.ts
@@ -1,0 +1,5 @@
+import { convertNxExecutor } from '@nrwl/devkit';
+
+import lintExecutor from './lint.impl';
+
+export default convertNxExecutor(lintExecutor);
