+++ packages/linter/src/executors/lint/schema.d.ts
@@ -1,18 +1,18 @@
 export interface Schema {
   linter: 'eslint' | 'tslint';
   config: string;
-  tsConfig: string | string;
+  tsConfig?: string | string[];
   format: Formatter;
   exclude: string[];
   files: string[];
-  force: boolean;
-  silent: boolean;
-  fix: boolean;
-  cache: boolean;
-  outputFile: string;
-  cacheLocation: string;
+  force?: boolean;
+  silent?: boolean;
+  fix?: boolean;
+  cache?: boolean;
+  outputFile?: string;
+  cacheLocation?: string;
   maxWarnings: number;
-  quiet: boolean;
+  quiet?: boolean;
 }
 
 type Formatter =
