+++ packages/linter/src/executors/lint/schema.json
@@ -2,6 +2,7 @@
   "$schema": "http://json-schema.org/schema",
   "title": "Lint Target",
   "description": "Linter",
+  "cli": "nx",
   "type": "object",
   "properties": {
     "linter": {
