+++ packages/linter/src/executors/lint/utility/file-utils.spec.ts
